TT-63: Create description screen presenter
