+++ 9.1_TikTalk_mobile/Podcast/Description/Presenter/DescriptionPresenter.swift
@@ -1,0 +1,20 @@
+import Foundation
+
+
+final class DescriptionPresenter {
+    
+    weak var viewController: DescriptionViewController?
+    private var description: Description {
+        didSet {
+            getDescription()
+        }
+    }
+    
+    init(description: Description) {
+        self.description = description
+    }
+    
+    func getDescription() {
+        viewController?.config(description)
+    }
+}
