+++ 9.1_TikTalk_mobile/Podcast/Description/View/DescriptionViewController.swift
@@ -1,0 +1,75 @@
+import UIKit
+
+
+final class DescriptionViewController: UIViewController {
+    
+    private var presenter: DescriptionPresenter
+    
+    private lazy var nameLabel: UILabel = {
+        let label = UILabel()
+        label.textColor = .white
+        label.font = .systemFont(ofSize: 20, weight: .medium)
+        label.textAlignment = .center
+        return label
+    }()
+    
+    private lazy var descriptionText: UITextView = {
+        let textView = UITextView()
+        textView.backgroundColor = .clear
+        textView.font = .systemFont(ofSize: 16, weight: .regular)
+        return textView
+    }()
+    
+    init(presenter: DescriptionPresenter) {
+        self.presenter = presenter
+        super.init(nibName: nil, bundle: nil)
+    }
+    
+    required init?(coder: NSCoder) {
+        fatalError("init(coder:) has not been implemented")
+    }
+    
+    override func viewDidLoad() {
+        super.viewDidLoad()
+        setup()
+    }
+    
+    func config(_ description: Description) {
+        nameLabel.text = description.podcastName
+        descriptionText.text = description.description
+    }
+}
+
+private extension DescriptionViewController {
+    
+    func setup() {
+        setupAppearance()
+        addSubviews()
+        activateConstraints()
+        presenter.getDescription()
+    }
+    
+    func setupAppearance() {
+        view.backgroundColor = UIColor(named: "Background")
+    }
+    
+    func addSubviews() {
+        [nameLabel,
+         descriptionText].forEach{
+            $0.translatesAutoresizingMaskIntoConstraints = false
+            view.addSubview($0)
+        }
+    }
+    
+    func activateConstraints() {
+        NSLayoutConstraint.activate([
+            nameLabel.topAnchor.constraint(equalTo: view.safeAreaLayoutGuide.topAnchor, constant: 10),
+            nameLabel.leadingAnchor.constraint(equalTo: view.safeAreaLayoutGuide.leadingAnchor, constant: 40),
+            nameLabel.trailingAnchor.constraint(equalTo: view.safeAreaLayoutGuide.trailingAnchor, constant: -40),
+            descriptionText.topAnchor.constraint(equalTo: nameLabel.bottomAnchor, constant: 20),
+            descriptionText.leadingAnchor.constraint(equalTo: view.safeAreaLayoutGuide.leadingAnchor, constant: 20),
+            descriptionText.trailingAnchor.constraint(equalTo: view.safeAreaLayoutGuide.trailingAnchor, constant: -20),
+            descriptionText.bottomAnchor.constraint(equalTo: view.safeAreaLayoutGuide.bottomAnchor)
+        ])
+    }
+}
