+++ 9.1_TikTalk_mobile/Podcast/Presenter/PodcastRouter.swift
@@ -4,7 +4,11 @@ import UIKit
 final class PodcastRouter {
     
     func showDescriptionFrom(_ viewController: UIViewController, description: Description) {
-        
+        let presenter = DescriptionPresenter(description: description)
+        let descriptionViewController = DescriptionViewController(presenter: presenter)
+        presenter.viewController = descriptionViewController
+        descriptionViewController.modalPresentationStyle = .formSheet
+        viewController.present(descriptionViewController, animated: true)
     }
     
     func showAuthorFrom(_ viewController: UIViewController, author: AuthorModel) {
